patient_triage/requirements.txt
pydantic>=2.7.0
openai>=1.40.0
openenv>=0.1.0
python-dotenv>=1.0.1
